restructured client to use react-router. The app now routes to pages depending on the path. Pages will serve specific aspects of our app and will utilize components. Adjusted some of the components.
