+++ client/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-//import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import {
   ApolloClient,
   InMemoryCache,
@@ -8,9 +8,12 @@ import {
 } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 import Header from './components/Header';
-import Body from './components/Body';
 import Footer from './components/Footer';
 import SignUpForm from './pages/SignUpForm'
+import LoginForm from './pages/LoginForm'
+import Profile from './pages/Profile'
+import Game from './pages/Game'
+import Home from './pages/Home'
 
 const httpLink = createHttpLink({
   uri: '/graphql',
@@ -34,10 +37,32 @@ const client = new ApolloClient({
 function App() {
   return (
     <ApolloProvider client={client}>
-      {/* <Header/> */}
-      <Body/>
-      {/* <Footer/> */}
-      {/* <SignUpForm/> */}
+      <Router>
+        {/* <Header/> */}
+        <Routes>
+          <Route 
+            path='/'
+            element={<Home />}
+          />
+          <Route 
+            path="/profile" 
+            element={<Profile />} 
+          />
+          <Route 
+            path="/signup" 
+            element={<SignUpForm />} 
+          />
+          <Route 
+            path="/login" 
+            element={<LoginForm />} 
+          />
+          <Route 
+            path='/game'
+            element={<Game />}
+          />
+        </Routes>
+        <Footer/>
+      </Router>
     </ApolloProvider>
   );
 }
