--- client/src/pages/Room.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function Room() {
-
-
-    return (
-        <div></div>
-    )
-}
